Renaming namespaces, cleaning up (#3)
