+++ src/Azure.Functions.Cli/Kubernetes/Models/IKubernetesResource.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Azure.Functions.Cli.Kubernetes.Models
+{
+    public interface IKubernetesResource
+    {
+    }
+}
