+++ src/Azure.Functions.Cli/Kubernetes/Models/IKubernetesSpec.cs
@@ -1,0 +1,6 @@
+namespace Azure.Functions.Cli.Kubernetes.Models
+{
+    public interface IKubernetesSpec
+    { 
+    }
+}
