+++ src/Azure.Functions.Cli/Kubernetes/Models/Kubernetes/BaseKubernetesResource.cs
@@ -3,10 +3,6 @@ using Newtonsoft.Json;
 
 namespace Azure.Functions.Cli.Kubernetes.Models.Kubernetes
 {
-    public interface IKubernetesResource
-    { 
-    }
-
     public abstract class BaseKubernetesResource<T> : IKubernetesResource where T : IKubernetesSpec
     {
         [JsonProperty("apiVersion")]
@@ -20,20 +16,5 @@ namespace Azure.Functions.Cli.Kubernetes.Models.Kubernetes
 
         [JsonProperty("spec")]
         public T Spec { get; set; }
-    }
-
-    public class SelectorV1
-    {
-        [JsonProperty("matchLabels")]
-        public IDictionary<string, string> MatchLabels { get; set; }
-    }
-
-    public class SearchResultV1<T>
-    {
-        [JsonProperty("apiVersion")]
-        public string ApiVersion { get; set; }
-
-        [JsonProperty("items")]
-        public IEnumerable<T> Items { get; set; }
     }
 }
