--- src/Azure.Functions.Cli/Kubernetes/Models/Kubernetes/IKubernetesSpec.cs
@@ -1,6 +1,0 @@
-namespace Azure.Functions.Cli.Kubernetes.Models.Kubernetes
-{
-    public interface IKubernetesSpec
-    { 
-    }
-}
