+++ src/Azure.Functions.Cli/Kubernetes/Models/Kubernetes/SearchResultV1.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Azure.Functions.Cli.Kubernetes.Models.Kubernetes
+{
+    public class SearchResultV1<T>
+    {
+        [JsonProperty("apiVersion")]
+        public string ApiVersion { get; set; }
+
+        [JsonProperty("items")]
+        public IEnumerable<T> Items { get; set; }
+    }
+}
