+++ src/Azure.Functions.Cli/Kubernetes/Models/Kubernetes/SelectorV1.cs
@@ -1,0 +1,11 @@
+﻿using Newtonsoft.Json;
+using System.Collections.Generic;
+
+namespace Azure.Functions.Cli.Kubernetes.Models.Kubernetes
+{
+    public class SelectorV1
+    {
+        [JsonProperty("matchLabels")]
+        public IDictionary<string, string> MatchLabels { get; set; }
+    }
+}
